pyproject.toml
[build-system]
requires = ["setuptools>=64.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "blackjack-simulator"
version = "0.1.0"
description = "A sophisticated Blackjack game simulator with multiple strategies"
readme = "README.md"
authors = [
    {name = "Ardeshir", email = "a.damavandi@hotmail.com"}
]
license = {text = "MIT"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Games/Entertainment",
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
requires-python = ">=3.7"
dependencies = [
    "pandas>=1.3.0",
    "numpy>=1.21.0",
    "matplotlib>=3.4.0",
    "openpyxl>=3.0.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    'black>=21.5b2',
    'flake8>=3.9.2',
]

[project.scripts]
blackjack = "src.game.blackjack:main" 